flux: created initial flux executor script
Problem: Current ATS functionality does not include scheduling for Flux,
the scheduler on El Capitan and other future LC machines. Expanded support
for this will allow users to begin to port testing to current and future
machines which use Flux as a system instance scheduler.

Note: currently this overrides the slurm56 flag to run on Ruby. This change
should not be included in the final merge or it will break compatibility on other
machines.

Fixes #88